功能优化，各个分类加入缓存，商品新增上下架，新增分类都能及时更新缓存，前端页面优化

+++ mailuser/src/main/java/com/example/mailuser/controller/ProductController.java
@@ -61,7 +61,7 @@ public class ProductController {
      */
     @GetMapping("/category/{categoryId}")
     @ApiOperation(value = "根据分类ID获取商品列表", notes = "根据选择的分类ID，显示该分类下的所有上架商品")
-    public Result<List<ProductVO>> getProductsByCategoryId(@PathVariable Integer categoryId) {
+    public Result<List<ProductVO>> getProductsByCategoryId(@PathVariable Long categoryId) {
         List<ProductVO> products = productService.getProductsByCategoryId(categoryId);
         return Result.success(products);
     }
